Refactor Argilla client structure and improve module organization
- Moved the Argilla client implementation to a new core module for better separation of concerns.
- Created a resources module to manage collections of Argilla resources (Users, Workspaces, Datasets, Webhooks).
- Updated the main client initialization to use the new structure and maintain backward compatibility.
- Improved documentation and comments for clarity.
- Removed circular import issues by restructuring imports.
